优化打包

+++ zipall.js
@@ -1,19 +1,19 @@
 const zipafolder = require('zip-a-folder')
 const fs = require('fs-extra')
 
-function plusVersion(version) {
-  const s = version.split('.')
-  return `${s[0]}.${s[1]}.${Number(s[2]) + 1}`
-}
-
 async function start() {
-  const version = fs.readFileSync('./version.txt').toString()
-  const newVersion = plusVersion(version)
-  fs.writeFileSync('./version.txt', newVersion)
-  console.log('开始压缩文件')
-  await zipafolder.zip(`./tempzip`, `./aurora-all-v${newVersion}.zip`)
-  console.log('压缩包已生成')
+  const config = fs.readJsonSync('./public/config.json')
+  if (!config.version) {
+    throw new Error('public/config.json is missing "version"')
+  }
+
+  console.log('Creating combined package')
+  await zipafolder.zip('./tempzip', `./aurora-all-v${config.version}.zip`)
+  console.log('Combined package generated')
   fs.removeSync('./tempzip')
 }
 
-start()
+start().catch((err) => {
+  console.error(err)
+  process.exit(1)
+})
